Add utils.c and basic functions
I added the file utils.c, which gets compiled and linked together with
game.c inside of the Makefile, and then proceeded to create and define
some functions inside of utils.c which I will use for the game.
Add utils.c and basic functions

I added the file utils.c, which gets compiled and linked together with
game.c inside of the Makefile, and then proceeded to create and define
some functions inside of utils.c which I will use for the game.