Finish basic game
I wrote the basic game, which by default has 6 obstacles, but that can
(and probably will eventually) be changed based on the number inside of
defs.h, but I will probably make it a const int that varied based off of
the screen size.

I added obstacles that can warp around the screen, and they vary in
size. The user's jumps are very smooth, and make sense. If the user hits
an obstacle, the game instantly dies, but I want to make the end screen
much better looking, so I will probably write a function to do that.

The game keeps track of your score in the top left-hand corner, and each
time you clear an obstacle, your score is incremented by 1.