Add 'install' and 'production' targets to Makefile
I added a 'production' target to the Makefile to make a stripped version
of the program, without debug_info or any Assembly. Not good for
debugging, but a smaller file size.

The 'install' target will make the 'production' target, and then copy it
into /usr/local/bin/, which will mean that is it on PATH, and is not an
executable file that you can run from anywhere in the terminal.